+++ clang-tools-extra/clang-tidy/misc/MiscTidyModule.cpp
@@ -22,6 +22,7 @@
 #include "NonCopyableObjects.h"
 #include "NonPrivateMemberVariablesInClassesCheck.h"
 #include "PersistenceStreamCheck.h"
+#include "PublicMembersCheck.h"
 #include "RedundantExpressionCheck.h"
 #include "StaticAssertCheck.h"
 #include "StructuredNamespaceCheck.h"
@@ -62,6 +63,8 @@ public:
         "misc-non-private-member-variables-in-classes");
     CheckFactories.registerCheck<PersistenceStreamCheck>(
         "misc-persistence-stream");
+    CheckFactories.registerCheck<PublicMembersCheck>(
+        "misc-public-members");
     CheckFactories.registerCheck<RedundantExpressionCheck>(
         "misc-redundant-expression");
     CheckFactories.registerCheck<StaticAssertCheck>("misc-static-assert");