+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -150,6 +150,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`misc-public-members
+  <clang-tidy/checks/misc/public-members>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`misc-structured-namespace
   <clang-tidy/checks/misc/structured-namespace>` check.